apps/mcp-servers/ocr/requirements.txt
mcp>=1.2.0
rapidocr-onnxruntime>=1.4.0
pillow>=10.0.0
uvicorn>=0.32.0
